requirements.txt
asgiref==3.4.1
autopep8
CacheControl==0.12.10
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.9
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
firebase-admin==5.1.0
gcloud==0.18.3
google-api-core==2.2.2
google-api-python-client==2.30.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-core==2.2.1
google-cloud-firestore==2.3.4
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpcio==1.41.1
grpcio-status==1.41.1
httplib2==0.20.2
idna==3.3
jws==0.1.3
msgpack==1.0.2
oauth2client==4.1.3
packaging==21.2
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle 
PyJWT==2.3.0
pyparsing==2.4.7
Pyrebase4==4.5.0
python-jwt==2.0.1
pytz==2021.3
requests==2.26.0
requests-toolbelt==0.9.1
rsa==4.7.2
six==1.16.0
sqlparse==0.4.2
toml
uritemplate==4.1.1
urllib3==1.26.7
wincertstore==0.2
gunicorn
django-heroku
tensorflow-cpu==2.8.0
pillow
django-cors-headers
django-sendgrid-v5
mediafire-dl @ git+https://github.com/Juvenal-Yescas/mediafire-dl@c73fe1e76632e6f9f51ed8167b4562ab1cfe03f7
